Adding Invariant test to the keygen, reshare, refresh and moving run_protocol functions to test
